Changing rendering so that SkillPills display next to the relevant JobCard on lg+ screens
